requirements.txt
python-telegram-bot==20.5  # Use your library version
